Bring in the dtc repo
The device tree compiler is an upstream project. However, we want to
be able to easily add patches which are not yet upstream, so a work-on
project is useful.

This will allow us to remove the patches that are currently applied
in the ebuild.

BUG=chromium-os:23249
TEST=manual: I'm sure there is a better way
$ mkdir /tmp/c
$ repo init -u https://git.chromium.org/chromiumos/manifest.git ...
$ copy my manifest file into .repo/manifest.xml
$ repo sync -j20 -n
$ ls /tmp/c/.repo/projects/src/third_party/dtc.git/
(see that it exists)

Change-Id: Ia56126f988198cbdf34ac4844c52dd3d1e308cd4
Reviewed-on: https://gerrit.chromium.org/gerrit/17520
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>