version: bump SDK version to 717.0.0
Bootstrapping the SDK without cached binary packages is currently broken
due to a previous update in portage-stable, removing serf slot 0. Older
SDK versions included serf:0 but now serf:1 is the only slot available.
As a first step to building a new SDK, an older SDK is used as the
"seed" and GCC is updated in that seed. Unfortunately through a complex
chain of dependencies GCC depends on one of our cros-workon packages,
perhaps baselayout which depends on git which depends on subversion
which depends on serf but with serf:0 no longer around that package
doesn't resolve properly and the update fails.

Version 717 was built after serf:0 was dropped and it just happened to
work because the binary package of serf:0 was cached. Update so SDK
builds work even without the old package cached. :)

Change-Id: Ie283eba5d98d9d73787f2efc0462ad9bfedb9edf