master: drop chromium fork of pyelftools
I can't find anywhere we use this. The fork is also dead (last commit
four years ago). The only thing that could use it would be pax-utils but
we don't have the python use flag, so it doesn't need it. We should use
gentoo's (up to date) upstream version if we do still want it.